# docs/architecture/api/openapi-skeleton.yaml
openapi: 3.1.0
info:
  title: Rekentafel API
  version: 1.0.0
  description: |
    Netherlands-first table QR split-pay platform API (MVP).
    Money fields are integer cents (EUR). Timestamps ISO-8601 UTC.
    Cross-ref entity dictionary: Bill, BillLine, AllocatableUnit, Participant,
    Allocation, CheckoutIntent, Payment, PaymentSession, DiningSession.
  contact:
    name: Rekentafel Engineering
servers:
  - url: https://api.rekentafel.nl/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /t/{restaurant_slug}/{table_code}:
    get:
      tags: [Guest — QR & Menu]
      summary: Resolve table QR — empty-table menu and context
      operationId: resolveTableQr
      parameters:
        - $ref: "#/components/parameters/RestaurantSlug"
        - $ref: "#/components/parameters/TableCode"
      responses:
        "200":
          description: Table context and menu (no bill unless payment active + token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableLandingResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /tables/{table_id}/service-signals:
    post:
      tags: [Guest — Service Signals]
      summary: Call server / ready to order
      operationId: createServiceSignal
      parameters:
        - $ref: "#/components/parameters/TableId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceSignalRequest"
      responses:
        "201":
          description: Signal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSignal"
        "429":
          $ref: "#/components/responses/RateLimited"
  /payment-sessions/join:
    post:
      tags: [Guest — Payment Session]
      summary: Join payment session with token or PIN
      operationId: joinPaymentSession
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinPaymentSessionRequest"
      responses:
        "200":
          description: Joined — returns participant credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinPaymentSessionResponse"
        "401":
          $ref: "#/components/responses/InvalidToken"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /payment-sessions/{payment_session_id}:
    get:
      tags: [Guest — Payment Session]
      summary: Get payment session state, bill summary, participants
      operationId: getPaymentSession
      security:
        - PaymentParticipantAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentSessionId"
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSessionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /payment-sessions/{payment_session_id}/events:
    get:
      tags: [Guest — Payment Session]
      summary: SSE stream for bill, claim, and payment updates
      operationId: subscribePaymentSessionEvents
      security:
        - PaymentParticipantAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentSessionId"
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
  /payment-sessions/{payment_session_id}/claims:
    get:
      tags: [Guest — Claims & Splits]
      summary: List allocations for session
      operationId: listClaims
      security:
        - PaymentParticipantAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentSessionId"
      responses:
        "200":
          description: Allocations
          content:
            application/json:
              schema:
                type: object
                properties:
                  bill_version:
                    type: integer
                  allocations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Allocation"
                  unclaimed_cents:
                    type: integer
                  remaining_cents:
                    type: integer
    post:
      tags: [Guest — Claims & Splits]
      summary: Create item or shared claim
      operationId: createClaim
      security:
        - PaymentParticipantAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentSessionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchBillVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimRequest"
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "423":
          $ref: "#/components/responses/Locked"
...
  /payment-sessions/{payment_session_id}/checkout:
    post:
      tags: [Guest — Checkout]
      summary: Initiate Mollie checkout for participant
      operationId: initiateCheckout
      security:
        - PaymentParticipantAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentSessionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchBillVersion"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Checkout created — redirect to Mollie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "409":
          description: Stale bill version or allocation conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "423":
          $ref: "#/components/responses/Locked"
...
  /staff/tables:
    get:
      tags: [Staff — Tables & Sessions]
      summary: List tables with session state
      operationId: listStaffTables
      security:
        - StaffAuth: []
      responses:
        "200":
          description: Table list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffTableSummary"
  /staff/tables/{table_id}/dining-sessions:
    post:
      tags: [Staff — Tables & Sessions]
      summary: Start dining session (Flow C)
      operationId: startDiningSession
      security:
        - StaffAuth: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                party_size:
                  type: integer
                  minimum: 1
                  maximum: 99
      responses:
        "201":
          description: Dining session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningSession"
...
  /staff/tables/{table_id}/bills:
    post:
      tags: [Staff — Bills]
      summary: Create or replace draft bill (manual entry)
      operationId: createBill
      security:
        - StaffAuth: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    get:
      tags: [Staff — Bills]
      summary: Get current bill for table
      operationId: getBill
      security:
        - StaffAuth: []
      parameters:
        - $ref: "#/components/parameters/TableId"
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
  /staff/tables/{table_id}/payment-sessions:
    post:
      tags: [Staff — Tables & Sessions]
      summary: Activate payment mode — issue PaymentSessionToken
      operationId: activatePaymentMode
      security:
        - StaffAuth: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_seconds:
                  type: integer
                  default: 7200
                  description: Default 2 hours
      responses:
        "201":
          description: Payment session opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivatePaymentResponse"
...
  /staff/tables/{table_id}/close:
    post:
      tags: [Staff — Tables & Sessions]
      summary: Close table — end session and reset
      operationId: closeTable
      security:
        - StaffAuth: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: X-Manager-PIN
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseTableRequest"
      responses:
        "200":
          description: Table closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseTableResponse"
  /staff/service-signals:
    get:
      tags: [Staff — Tables & Sessions]
      summary: Inbox of pending service signals
      operationId: listServiceSignals
      security:
        - StaffAuth: []
      responses:
        "200":
          description: Signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceSignal"
  /staff/service-signals/{signal_id}/ack:
    post:
      tags: [Staff — Tables & Sessions]
      summary: Acknowledge service signal
      operationId: ackServiceSignal
      security:
        - StaffAuth: []
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Acknowledged
  /webhooks/mollie:
    post:
      tags: [Webhooks]
      summary: Mollie payment status webhook ingress
      operationId: mollieWebhook
      description: |
        Returns 200 within 500ms. Processing is async.
        Verification via fetch payment API — not redirect URL.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Mollie payment id e.g. tr_xxx
      responses:
        "200":
          description: Accepted for processing
...
components:
  securitySchemes:
    GuestAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Guest device session JWT (typ=guest)
    PaymentParticipantAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Payment participant JWT (typ=payment_participant)
    StaffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff JWT with restaurant_id and role
  parameters:
    RestaurantSlug:
      name: restaurant_slug
      in: path
      required: true
      schema:
        type: string
        example: de-rekentafel
    TableCode:
      name: table_code
      in: path
      required: true
      schema:
        type: string
        example: T12
    TableId:
      name: table_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PaymentSessionId:
      name: payment_session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
    IfMatchBillVersion:
      name: If-Match
      in: header
      description: 'Format bill_version:N e.g. bill_version:12'
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Invalid or expired token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidToken:
      description: Invalid payment session token or PIN
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Concurrent modification or unit unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unprocessable:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Locked:
      description: Resource locked — claims frozen or checkout in progress
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          example: https://api.rekentafel.nl/errors/unit-unavailable
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          enum:
            - UNIT_UNAVAILABLE
            - UNIT_LOCKED
            - BILL_VERSION_STALE
            - CLAIMS_FROZEN
            - CHECKOUT_IN_PROGRESS
            - SHARE_OVERFLOW
            - PLEDGE_EXCEEDS_REMAINING
            - IDEMPOTENCY_KEY_REUSED
            - PIN_LOCKED
        retry_after_ms:
          type: integer
...
    TableSessionState:
      type: string
      enum: [EMPTY, SEATED, PAYMENT_ACTIVE, CLOSED]
...
    SplitMode:
      type: string
      enum: [ITEM, EQUAL, CUSTOM, SHARED]
...
    TableLandingResponse:
      type: object
      required: [table, restaurant, session_state, menu]
      properties:
        table:
          $ref: "#/components/schemas/Table"
        restaurant:
          type: object
          properties:
            restaurant_id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
        session_state:
          $ref: "#/components/schemas/TableSessionState"
        payment_session_hint:
          type: object
          nullable: true
          description: Present only if PAYMENT_ACTIVE — no bill without join
          properties:
            payment_session_id:
              type: string
              format: uuid
            join_required:
              type: boolean
              example: true
        menu:
          $ref: "#/components/schemas/Menu"
...
    CreateServiceSignalRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [ready_to_order, assistance]
    ServiceSignal:
      type: object
      properties:
        signal_id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [pending, acknowledged, expired]
        cooldown_until:
          type: string
          format: date-time
    DiningSession:
      type: object
      properties:
        dining_session_id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        waiter_id:
          type: string
          format: uuid
        party_size:
          type: integer
        started_at:
          type: string
          format: date-time
    Bill:
      type: object
      required: [bill_id, bill_version, lines, bill_grand_total_cents]
      properties:
        bill_id:
          type: string
          format: uuid
        dining_session_id:
          type: string
          format: uuid
        bill_version:
          type: integer
          minimum: 1
        currency:
          type: string
          enum: [EUR]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/BillLine"
        bill_grand_total_cents:
          $ref: "#/components/schemas/MoneyCents"
        service_charge_cents:
          $ref: "#/components/schemas/MoneyCents"
        settlement:
          $ref: "#/components/schemas/BillSettlement"
...
    CreateBillRequest:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          items:
            type: object
            required: [name, qty, unit_price_inc_vat_cents, vat_rate_bps]
            properties:
              name:
                type: string
              qty:
                type: integer
              unit_price_inc_vat_cents:
                $ref: "#/components/schemas/MoneyCents"
              vat_rate_bps:
                type: integer
              splittable:
                type: boolean
        apply_service_charge:
          type: boolean
          default: true
...
    JoinPaymentSessionRequest:
      type: object
      required: [payment_session_id, token, token_type, display_name]
      properties:
        payment_session_id:
          type: string
          format: uuid
        token:
          type: string
          description: Raw secret token or 6-digit join_pin
        token_type:
          type: string
          enum: [secret, pin]
        display_name:
          type: string
          minLength: 1
          maxLength: 32
    JoinPaymentSessionResponse:
      type: object
      required: [participant_id, participant_token, payment_session]
      properties:
        participant_id:
          type: string
          format: uuid
        participant_token:
          type: string
          description: JWT for payment_participant auth
        payment_session:
          $ref: "#/components/schemas/PaymentSessionDetail"
    PaymentSessionDetail:
      type: object
      properties:
        payment_session_id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        bill_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [OPEN, CLOSED, EXPIRED]
        expires_at:
          type: string
          format: date-time
        bill:
          $ref: "#/components/schemas/Bill"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        settlement:
          $ref: "#/components/schemas/BillSettlement"
...
    Allocation:
      type: object
      required: [allocation_id, participant_id, split_mode, amount_cents, bill_version]
      properties:
        allocation_id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        unit_id:
          type: string
          format: uuid
        split_mode:
          $ref: "#/components/schemas/SplitMode"
        share_numerator:
          type: integer
        share_denominator:
          type: integer
        amount_cents:
          $ref: "#/components/schemas/MoneyCents"
        bill_version:
          type: integer
        equal_group_id:
          type: string
          format: uuid
    CreateClaimRequest:
      type: object
      required: [split_mode]
      properties:
        split_mode:
          type: string
          enum: [ITEM, SHARED]
        unit_id:
          type: string
          format: uuid
        share_numerator:
          type: integer
        share_denominator:
          type: integer
          maximum: 20
...
    CheckoutRequest:
      type: object
      properties:
        tip_cents:
          $ref: "#/components/schemas/MoneyCents"
        redirect_url:
          type: string
          format: uri
          description: Override default return URL
    CheckoutResponse:
      type: object
      required: [checkout_id, mollie_checkout_url, total_cents, expires_at]
      properties:
        checkout_id:
          type: string
          format: uuid
        mollie_payment_id:
          type: string
          example: tr_abc123
        mollie_checkout_url:
          type: string
          format: uri
        food_cents:
          $ref: "#/components/schemas/MoneyCents"
        tip_cents:
          $ref: "#/components/schemas/MoneyCents"
        total_cents:
          $ref: "#/components/schemas/MoneyCents"
        expires_at:
          type: string
          format: date-time
...
    StaffTableSummary:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/Table"
        dining_session:
          $ref: "#/components/schemas/DiningSession"
        payment_session_id:
          type: string
          format: uuid
        settlement:
          $ref: "#/components/schemas/BillSettlement"
        pending_signals:
          type: integer
...
    CloseTableRequest:
      type: object
      required: [close_reason]
      properties:
        close_reason:
          type: string
          enum: [FULLY_PAID, FORCE_CLOSE, CASH_REMAINDER, EXTERNAL_POS, WALKOUT]
        external_payment_note:
          type: string
...